+++ engine/src/main/java/net/jqwik/engine/properties/arbitraries/IntegralEdgeCasesConfiguration.java
@@ -1,8 +1,6 @@
 package net.jqwik.engine.properties.arbitraries;
 
 import java.math.*;
-import java.util.*;
-import java.util.function.*;
 
 import net.jqwik.api.*;
 import net.jqwik.engine.properties.*;
@@ -16,6 +14,14 @@ class IntegralEdgeCasesConfiguration extends GenericEdgeCasesConfiguration<BigIn
 	public IntegralEdgeCasesConfiguration(Range<BigInteger> range, BigInteger shrinkingTarget) {
 		this.range = range;
 		this.shrinkingTarget = shrinkingTarget;
+	}
+
+	@Override
+	protected void checkEdgeCaseIsValid(BigInteger edgeCase) {
+		if (!range.includes(edgeCase)) {
+			String message = String.format("Edge case <%s> is outside the arbitrary's allowed range %s", edgeCase, range);
+			throw new IllegalArgumentException(message);
+		}
 	}
 
 	@Override